website files

+++ operation.php
@@ -3,13 +3,17 @@ session_start();
 require_once "connect.php";
 require_once "function.php";
 
+if(isset($_SESSION["user"])&&$_SESSION["user"]!=""){
+$id=$_SESSION["user"];
+}
+
 if($_GET["operation"]=="add"){
     $prod=$_GET["prod"];
-    $sql = "SELECT * from cart where id=$prod";
+    $sql = "SELECT * from cart where id=$prod and userid=$id";
     $result=$conn -> query($sql);
     $num_row=mysqli_num_rows($result);
     if($num_row==0){
-        $sql = "INSERT INTO cart(id, amount) values($prod, 1)";
+        $sql = "INSERT INTO cart(id, amount, userid) values($prod, 1, $id)";
         $result=$conn->query($sql);
         $sql = " SELECT nazwa FROM produkty as p JOIN type as t ON p.type = t.id WHERE p.id=$prod";
         $result=$conn->query($sql);
@@ -23,7 +27,7 @@ if($_GET["operation"]=="add"){
         $row = $result -> fetch_assoc();
         if($row['amount']<5){
             $amount = $row['amount']+1;
-            $sql = "UPDATE cart SET amount = $amount";
+            $sql = "UPDATE cart SET amount = $amount where userid=$id";
             $result = $conn->query($sql);
             $sql = " SELECT nazwa FROM produkty as p JOIN type as t ON p.type = t.id WHERE p.id=$prod";
             $result=$conn->query($sql);
@@ -44,11 +48,11 @@ if($_GET["operation"]=="add"){
 
 if($_GET["operation"]=="addindex"){
     $prod=$_GET["prod"];
-    $sql = "SELECT * from cart where id=$prod";
+    $sql = "SELECT * from cart where id=$prod and userid=$id";
     $result=$conn -> query($sql);
     $num_row=mysqli_num_rows($result);
     if($num_row==0){
-        $sql = "INSERT INTO cart(id, amount) values($prod, 1)";
+        $sql = "INSERT INTO cart(id, amount, userid) values($prod, 1, $id)";
         if($result=$conn->query($sql)){
             header("Location:index.php?add=yes");
         }
@@ -59,7 +63,7 @@ if($_GET["operation"]=="addindex"){
         $row = $result -> fetch_assoc();
         if($row['amount']<5){
             $amount = $row['amount']+1;
-            $sql = "UPDATE cart SET amount = $amount";
+            $sql = "UPDATE cart SET amount = $amount where userid=$id";
             $result = $conn->query($sql);
             header("Location:index.php?add=yes");
         }
@@ -72,11 +76,11 @@ if($_GET["operation"]=="addindex"){
 
 if($_GET["operation"]=="addspec"){
     $prod=$_GET["prod"];
-    $sql = "SELECT * from cart where id=$prod";
+    $sql = "SELECT * from cart where id=$prod and userid=$id";
     $result=$conn -> query($sql);
     $num_row=mysqli_num_rows($result);
     if($num_row==0){
-        $sql = "INSERT INTO cart(id, amount) values($prod, 1)";
+        $sql = "INSERT INTO cart(id, amount, userid) values($prod, 1, $id)";
         if($result=$conn->query($sql)){
             header("Location:spec.php?prod=$prod&add=yes");
         }
@@ -87,7 +91,7 @@ if($_GET["operation"]=="addspec"){
         $row = $result -> fetch_assoc();
         if($row['amount']<5){
             $amount = $row['amount']+1;
-            $sql = "UPDATE cart SET amount = $amount";
+            $sql = "UPDATE cart SET amount = $amount where userid=$id";
             $result = $conn->query($sql);
             header("Location:spec.php?prod=$prod&add=yes");
         }
@@ -101,7 +105,7 @@ if($_GET["operation"]=="addspec"){
 
 if($_GET["operation"]=="del"){
     $prod=$_GET["prod"];
-    $sql = "DELETE FROM cart where id=$prod";
+    $sql = "DELETE FROM cart where id=$prod and userid=$id";
     $result=$conn->query($sql);
     header("Location:cart.php");
 }
@@ -109,7 +113,7 @@ if($_GET["operation"]=="del"){
 if($_GET["operation"]=="amount"){
     $amount=$_POST["amount"];
     $prod=$_GET["prod"];
-    $sql = "UPDATE cart SET amount=$amount where id=$prod";
+    $sql = "UPDATE cart SET amount=$amount where id=$prod and userid=$id";
     $result=$conn->query($sql);
     header("Location:cart.php");
 }
@@ -117,6 +121,22 @@ if($_GET["operation"]=="amount"){
 if($_GET["operation"]=="logout"){
     $_SESSION["user"]="";
     header('Location:index.php');
+}
+
+if($_GET["operation"]=="logindex"){
+    header('Location:index.php?log=yes');
+}
+if($_GET["operation"]=="logspec"){
+    $prod=$_GET["prod"];
+    header("Location:spec.php?prod=$prod&log=yes");
+}
+if($_GET["operation"]=="logadd"){
+    $prod=$_GET["prod"];
+    $sql = " SELECT nazwa FROM produkty as p JOIN type as t ON p.type = t.id WHERE p.id=$prod";
+    $result=$conn->query($sql);
+    $row=$result->fetch_assoc();
+    $category=$row["nazwa"];
+    header("Location:ofert.php?category=$category&log=yes");
 }
 
 if((empty($_GET["operation"]))){
